requirements.txt
pyobjc-framework-Quartz
pyobjc-framework-Vision
python-dateutil
